requirements.txt
fastapi
psycopg2-binary
python-dotenv
uvicorn
python-multipart
pandas
streamlit